Adding concurrency field to model
