+++ docker-compose.yml
@@ -56,6 +56,8 @@ services:
   postgres:
     image: postgres
     restart: always
+    ports:
+      - 5432:5432
     environment:
       POSTGRES_USER: admin
       POSTGRES_PASSWORD: admin
